requirements.txt
flask==3.0.3
jinja2==3.1.4
paho-mqtt==2.1.0
pyyaml==6.0.2
apscheduler==3.10.4
SQLAlchemy>=2.0